requirements.txt
absl-py==1.4.0
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
arch==5.3.1
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
biopython==1.81
bleach==6.0.0
blessed==1.20.0
blis==0.7.10
build==1.0.0
cachetools==5.3.1
caldav==1.2.1
catalogue==2.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmaes==0.10.0
colorlog==6.7.0
colour==0.1.5
comm==0.1.2
confection==0.1.1
contourpy==1.0.7
cramjam==2.6.2
croniter==1.4.1
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
dash==2.9.2
dash-bio==1.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.3.0
dateparser==1.1.8
dateutils==0.6.12
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.4.1
dill==0.3.6
exceptiongroup==1.1.2
executing==1.2.0
fastai==2.7.12
fastapi==0.103.1
fastcore==1.5.29
fastdownload==0.0.7
fastparquet==2023.2.0
fastprogress==1.0.3
filelock==3.12.3
Flask==2.2.3
flatbuffers==23.5.26
fonttools==4.38.0
frozendict==2.3.8
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
GEOparse==2.0.3
GooCalendar==0.8.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
hottbox==0.3.2
html5lib==1.1
icalendar==5.0.7
idna==3.4
imageio==2.31.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
inquirer==3.1.3
ipykernel==6.21.2
ipython==8.11.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyterlab-widgets==3.0.5
kaleido==0.2.1
keras==2.13.1
kiwisolver==1.4.4
kneed==0.8.3
langcodes==3.3.0
libclang==16.0.6
lightning==2.0.8
lightning-cloud==0.5.37
lightning-utilities==0.9.0
llvmlite==0.40.1
locket==1.0.0
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
multitasking==0.0.11
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
optuna==3.3.0
ordered-set==4.1.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pandas-datareader==0.10.0
ParmEd==4.1.0
parso==0.8.3
partd==1.3.0
pathy==0.10.2
patsy==0.5.3
periodictable==1.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.0
plotly==5.13.0
preshed==3.0.8
prompt-toolkit==3.0.38
property-cached==1.6.4
protobuf==4.24.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.5.9
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic-core==2.4.0
pyfinance==1.3.0
Pygments==2.14.0
pyinstrument==4.4.0
PyJWT==2.8.0
pyparsing==3.0.9
pyproject-hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-forecasting==1.0.0
pytorch-ignite==0.4.12
pytorch-lightning==2.0.8
pytorch-optimizer==2.11.2
pytorchtools==0.0.2
pyts==0.13.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.0
ray==2.6.3
readchar==4.0.5
recurring-ical-events==2.0.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
schedule==1.2.0
scikit-learn==1.2.1
scikit-optimize==0.9.0
scipy==1.10.0
seaborn==0.12.2
selenium==4.10.0
six==1.16.0
sklearn==0.0
smart-open==6.3.0
snakeviz==2.1.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.20
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
statsmodels==0.13.5
swifter==1.3.4
sympy==1.12
tabulate==0.9.0
tardis-dev==2.0.0a13
tenacity==8.2.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
tensorly==0.8.1
termcolor==2.3.0
thinc==8.1.12
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchmetrics==1.1.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
uvicorn==0.23.2
vobject==0.9.6.1
wasabi==1.1.2
wcwidth==0.2.6
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
wsproto==1.2.0
x-wr-timezone==0.0.5
xgboost==1.7.5
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.9.2
yfinance==0.2.18
zipp==3.15.0
